Split src and include
